# tpl/partials: Fix recently introduced deadlock in partials cache