--- a/tpl/internal/go_templates/texttemplate/hugo_template_test.go
+++ b/tpl/internal/go_templates/texttemplate/hugo_template_test.go
@@ -1,4 +1,4 @@
-// Copyright 2019 The Hugo Authors. All rights reserved.
+// Copyright 2022 The Hugo Authors. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,6 +15,7 @@ package template
 
 import (
 	"bytes"
+	"context"
 	"reflect"
 	"strings"
 	"testing"
@@ -35,24 +36,26 @@ func (t TestStruct) Hello2(arg1, arg2 string) string {
 	return arg1 + " " + arg2
 }
 
-type execHelper struct {
+type execHelper struct{}
+
+func (e *execHelper) Init(ctx context.Context, tmpl Preparer) {
 }
 
-func (e *execHelper) GetFunc(tmpl Preparer, name string) (reflect.Value, bool) {
+func (e *execHelper) GetFunc(ctx context.Context, tmpl Preparer, name string) (reflect.Value, reflect.Value, bool) {
 	if name == "print" {
-		return zero, false
+		return zero, zero, false
 	}
 	return reflect.ValueOf(func(s string) string {
 		return "hello " + s
-	}), true
+	}), zero, true
 }
 
-func (e *execHelper) GetMapValue(tmpl Preparer, m, key reflect.Value) (reflect.Value, bool) {
+func (e *execHelper) GetMapValue(ctx context.Context, tmpl Preparer, m, key reflect.Value) (reflect.Value, bool) {
 	key = reflect.ValueOf(strings.ToLower(key.String()))
 	return m.MapIndex(key), true
 }
 
-func (e *execHelper) GetMethod(tmpl Preparer, receiver reflect.Value, name string) (method reflect.Value, firstArg reflect.Value) {
+func (e *execHelper) GetMethod(ctx context.Context, tmpl Preparer, receiver reflect.Value, name string) (method reflect.Value, firstArg reflect.Value) {
 	if name != "Hello1" {
 		return zero, zero
 	}
@@ -78,12 +81,11 @@ Method: {{ .Hello1 "v1" }}
 	var b bytes.Buffer
 	data := TestStruct{S: "sv", M: map[string]string{"a": "av"}}
 
-	c.Assert(ex.Execute(templ, &b, data), qt.IsNil)
+	c.Assert(ex.ExecuteWithContext(context.Background(), templ, &b, data), qt.IsNil)
 	got := b.String()
 
 	c.Assert(got, qt.Contains, "foo")
 	c.Assert(got, qt.Contains, "hello hugo")
 	c.Assert(got, qt.Contains, "Map: av")
 	c.Assert(got, qt.Contains, "Method: v2 v1")
-
 }